pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "reagent"
version = "0.1.0"
description = "Manage, secure, and evolve Claude Code assets across repositories"
readme = "README.md"
license = "MIT"
requires-python = ">=3.13"
authors = [
    { name = "calghar" },
]
keywords = ["claude-code", "ai-agents", "devtools", "asset-management"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "rich>=13.0",
    "jsonschema>=4.20",
    "httpx>=0.27",
    "jinja2>=3.1",
]

[project.optional-dependencies]
code-intel = ["mcp>=1.0"]
dashboard = ["starlette>=0.40", "uvicorn>=0.30", "aiosqlite>=0.20"]
docs = [
    "mkdocs-material>=9.5",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.8",
    "mypy>=1.10",
    "types-PyYAML>=6.0",
    "bandit>=1.7",
    "pip-audit>=2.7",
    "pre-commit>=3.7",
]

[project.scripts]
reagent = "reagent.cli:main"

[project.urls]
Homepage = "https://github.com/calghar/reagent"
Documentation = "https://calghar.github.io/reagent"
Repository = "https://github.com/calghar/reagent"
Issues = "https://github.com/calghar/reagent/issues"
Changelog = "https://github.com/calghar/reagent/blob/main/CHANGELOG.md"

[tool.hatch.build.targets.wheel]
packages = ["src/reagent"]

[tool.ruff]
target-version = "py313"
line-length = 88
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "S", "B", "A", "C4", "PTH"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.mypy]
python_version = "3.13"
strict = true
plugins = ["pydantic.mypy"]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[[tool.mypy.overrides]]
module = ["mcp", "mcp.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["aiosqlite", "aiosqlite.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["starlette", "starlette.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["uvicorn", "uvicorn.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["jsonschema", "jsonschema.*"]
ignore_missing_imports = true

[tool.bandit]
exclude_dirs = ["tests", "scripts"]
skips = ["B110", "B311", "B404", "B603", "B607", "B608"]

[tool.coverage.run]
source = ["reagent"]

[tool.coverage.report]
show_missing = true
skip_empty = true
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.",
]

[dependency-groups]
dev = [
    "pytest-mock>=3.15.1",
]
